move any functions that required importing a global journalList variable back into main.js because JS doesn't let you import multiple of the same variable (if in the future, we were to make list.js use util.js)
